pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-post"
version = "0.1.0"
description = "Instagram/X auto-posting system for woodcarving class photos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "gspread>=5.10.0",
    "boto3>=1.28.0",
    "tweepy>=4.14.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
auto-post = "auto_post.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_post"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
